backend/requirements.txt
langchain-ollama>=0.1.0
langchain>=0.2.0
langchain-core>=0.2.0
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
rank_bm25>=0.2.2
rouge-score>=0.1.2
transformers>=4.41.0
torch>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pypdf>=4.2.0
python-docx>=1.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
tenacity>=8.3.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
